app_test/requirements.txt
Appium-Python-Client>=2.0.0
pytest>=7.0.0
pytest-html>=3.0.0
allure-pytest>=2.9.0
PyYAML>=6.0
selenium>=4.0.0
pillow>=8.0.0
python-dotenv>=0.19.0
assertpy>=1.1
